divided VideoPlayer w/controls & VideoPlayerCustomyzed
