+++ src/Components/VideoPlayer/VideoPlayer.scss
@@ -76,7 +76,7 @@ video {
   display: flex;
   justify-content: space-around;
   width: 100%;
-  margin: 1rem 0;
+  margin: 2rem 0;
 
   & button {
     display: flex;
